Ruleset: selectively exclude an errorcode
As of PHPCS 4.0, it is possible to selectively exclude errors from the `PSR12.Files.FileHeader` sniff.

This particular exclusion means that we will not enforce a blank line between the PHP open tag and the file docblock, as that would cause a lot of unnecessary code churn in the primary codebases using this ruleset.