Composer/ruleset: update for PHPCompatibility 10.0.0
Already allow for PHPCompatibility 10.0.0 which is due in not too long.

For external standards which use this ruleset, always allow for the current major and the next major if in active development.
This prevents issues with Composer not being able to install on dev branches.

Includes some updates to the ruleset exclusions for new tokens being sniffed for in PHPCompatibility 10.0.0 (which are backfilled in PHPCS, so not a problem for PHPCS standards to use).