Composer/ruleset: allow for more PHP 8.0 token constants
PHP 8.0 will introduce yet more new token constants. A PR to sniff some more of those will be opened shortly in PHPCompatibility, so we should allow for these to be used in external PHPCS standards.

The token constants are expected to be backfilled by PHPCS as of version 3.5.7. The backfilling of the tokenization _using these tokens_ won't go into PHPCS until PHPCS 4.0.0.

See: squizlabs/PHP_CodeSniffer 3041