Ruleset: prevent sniff conflict
PHPCS 3.6.2 introduced a new sniff to the PSR12 standard, which enforces no blank line at the start of a class.

That sniff conflicts with the `Squiz.WhiteSpace.FunctionSpacing` sniff when the `spacingBeforeFirst` property is set to `1`.

As the property setting predates the new sniff, excluding the new sniff will prevent a lot of unnecessary code churn in repos using this standard.

This does, however, require raising the minimum PHPCS version to PHPCS 3.6.2 as otherwise a "sniff not found" error could be displayed.

Ref: https://github.com/squizlabs/PHP_CodeSniffer/releases/tag/3.6.2